+++ packages/eyes-api/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- add `LegacyRegion` type to support regions with `left` and `top` instead of `x` and `y`
+- updated to @applitools/types@1.0.14 (from 1.0.13)
 
 ## 1.1.1 - 2021/9/9